Insertion of several new functions, deletion and rearragement of older functions.
New functions :
- rescale_data : need to rescale the data (to positive values). This
data contain the position of the biomocules which has been reduced into
 lower dimension.
Negative values have been generated. By computing and summing up the
 computing and summing up the absolute error (abs_diff = pred - true),
this latter has significantly low values. The summation of negative and
 positive values, cancel the error.

- goal_prediction :  we are working on unsupervised data, no reference
  exist to measure our prediction to real values. This function will
provide reference, that correspond to the mean point of each cluster.
The points within each cluster should be close to the reference.

- cluster_sse : compute the sum squared error and the absolute error, in
order to evaluate the distance  between the reference data (true) and the real data.

- process_data : execute the funczions that load and transform the data,
  into lower dimension, then into reachability values.

- perform_rhc : to perform the clustering

- single_rhc : used if want to perform only one rhc

- deep_rhc : by executing this single function, the clustering is
  perform and optimized over iteration.

Deleted function :
- executed_revised_clustering : this has been separted in 2 functions :
  process_data and perform_rhc.

Rearranged function :
- dim_reduc : The plotting script has been removed for clarity purpose.
It will be in the future integrate in a new function