Modification in the function deep_rhcc : Errors were met if the initialization of the cutoff_min is too low close to 0, or too close to the maximum dist_reach. Change the interval value when chosing an initial cutoff_min randomly.
New function dropout were added :
To avoid overfitting in (potential) large data, preventing highlight of
some noise, accelerate the computation, a function was added to froze some
inputs nodes to 0.

The dropout function is called in gradient_sse function.
To compensate the masked node (here 30% of the data), we multiply
the "reduced data" by a factor.
The gradient formula (in the function gradient_sse) was modified.