Add function to subsample data.
As default, the data will not be subsampled.
However for very large data, it is recommended to subsample data.
If you have enough large data, subsampling it will not modify the
consistency of the data and it will improve the efficiency of the
algorithm.

Add dropout function.
In order to increase the efficiency of the data, as default 30% of data
will be dropout when optimizing the weight.
The dropout also allow to avoid overfitting.

Add return_features parameters.
It select the coordinates of the protein to be written in output xtc file.
As default, return_feature="protein", it will select all protein(s) from the input pdb
to be written in the output trajectory files.

In the function label,
A sorting function is applied, to write the labeling of each data point in ascending
number according the size of the cluster.

To improve the efficiency of this algorithm, I would like to compute the
hierarchichal linkage in parallel.
When I test the RHCC algorithm on very large data, the linkage algorithm
from the scipy package failed. It sends an error related to the memory
allocation.
A function calculate_linkage_multiprocess is added, but not yet
integrate to be used.
After computing my data in parallel, I failed to concatenate the
splitted data because there are overlapping clusters IDs.