Added a flag namely tag_iteration, which indicate if the iteration should pursue ("continue") or stopped ("stop").
The reason, was, when I worked on small data and used a large value of min_number_data, the script continue to
iterate eventough min_number_data is reached. During the iteration, seek_ID increase, but no new clusters were registed.
This has generated error, when it fetch a value from the indexes